Entities load values from overview pages automatically
This is the first iteration of a feature that will give callers access
to all data that's available in "overviews" of REST entities.

Each entity has usually both an overview REST endpoint as well as a details
endpoint. An example is the endpoint `configuration/workflows/<id>`,
which contains an overview of both protocols and stages.The behaviour
of this library was to fetch the details view to see those values,
so the following code would incur four calls to the server:

```
workflow = lims.get_workflows(name)[0]
print(workflow.stages[0].name)
print(workflow.protocols[0].name)

```

With this change, only the first two requests are executed.

This commit only solves the case for this particular use case and is a
hotfix. A commit will follow which cleans this up and provides the
feature for all use cases, along with integration tests.