Support batch fetching the same artifact in differnt state
Context: Artifacts can be fetched in a particular state by specifying
the `state` query parameter. Specifying no `state` parameter when
fetching artifacts will fetch it in the latest state, but Clarity
generally (I don't know of any exceptions) refers to Artifacts in a
particular state, e.g. in the input/output map for a process as well as
in the response when batch fetching Artifacts.

Patch:

The batch function was keeping a dict of all fetched entities by lims ID
rather than URI. This meant that it wasn't possible to query for the
same artifact in a different state. This patch fixes that by saving the
entries by URI instead of the ID.

Furthermore, if a user batch fetches the entries without a state
query parameter we will get entries *with* the state parameter. We have
to make sure that we map it to the correct requested Artifact in this case.